fix: Aspiration window overflowing Bench: 13971862

+++ engine/src/search/mod.rs
@@ -1072,7 +1072,7 @@ impl Search {
                     // -EvalNumber::MAX = -2147483647
                     // EvalNumber::MIN = -2147483648
 
-                    beta = (alpha + beta) / 2;
+                    beta = ((i64::from(alpha) + i64::from(beta)) / 2) as i32;
                 } else if best_score >= beta {
                     beta = beta.saturating_add(param!(self).aspiration_window_growth);
                 } else {
